Add tentative WPT to verify download in sandbox
General testing idea:
For a network request, we store a token in stash. And after a fixed
period of time sufficient for the download to finish, if there are
any, we validate the token in the stash to see if the network request
was issued and finished successfully.

In the case of <a download> where the decision of download can be made
before resource fetching, the server sets the token immediately.

In the case of navigation to a download, the server will stream a
response over 1 seconds and set the token at the end only when the
socket has been connected. So it is able to detect any download
cancellation while fetching the resource.

Bug: 539938
Change-Id: I734c8cc18d1f761f16646c6b859a6b731ab40ff5