Update cached azimuth/elevation/cone gain
Sample accurate panner automations need to update the cached value for
the azimuth, elevation, and cone gain at the end of each render in
case the automations end (because all automations are sufficiently far
in the past).  When this happens, the panner switches from a-rate
sample-accurate method to k-rate method that uses the cached values
when nothing is moving.  If the a-rate method doesn't update, then the
k-rate method uses the old values which are now incorrect.

Bug: 917757
Test: the-panner-node/automation-changes.html
Change-Id: I7289c278b1cc9dc31d9e3f5698ac467cb1474ca5